# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dreamsdesign Sales War Machine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments/slots:
    get:
      operationId: getAvailableSlots
      tags: [Appointments]
      summary: Get available time slots for a date
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        time:
                          type: string
                        label:
                          type: string
                      required: [time, label]
                required: [date, slots]
  /appointments:
    get:
      operationId: listAppointments
      tags: [Appointments]
      summary: List all appointments
      responses:
        "200":
          description: All appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [Appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Slot no longer available
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [Appointments]
      summary: Update appointment status or notes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [confirmed, cancelled, completed, noshow]
                notes:
                  type: string
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [Appointments]
      summary: Delete an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ─────────────────────────────────────────
  # WHATSAPP
  # ─────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        durationMinutes:
          type: integer
        timezone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        businessSummary:
          type: ["string", "null"]
        specificProblem:
          type: ["string", "null"]
        desiredResult:
          type: ["string", "null"]
        whyCanHelp:
          type: ["string", "null"]
        investmentWillingness:
          type: ["string", "null"]
        minInvestmentConfirm:
          type: ["string", "null"]
        startSoon:
          type: ["string", "null"]
        businessPartner:
          type: ["string", "null"]
        services:
          type: array
          items:
            type: string
        otherService:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, scheduledDate, scheduledTime, durationMinutes, status, createdAt, updatedAt]
    CreateAppointmentInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        location:
          type: string
          enum: [meet, inperson]
        businessSummary:
          type: string
        specificProblem:
          type: string
        desiredResult:
          type: string
        whyCanHelp:
          type: string
        investmentWillingness:
          type: string
        minInvestmentConfirm:
          type: string
        startSoon:
          type: string
        businessPartner:
          type: string
        services:
          type: array
          items:
            type: string
        otherService:
          type: string
      required: [name, email, scheduledDate, scheduledTime]

    # ─── WhatsApp Schemas ────────────────────────────────────────────
...